Adding Screenshots and add protection for cart

+++ controllers/cartController.js
@@ -1,5 +1,7 @@
+// Importing Files and Packages
 const Cart = require("../model/Cart");
 const Product = require("../model/Product");
+const mongoose = require("mongoose");
 
 // POST /cart
 const addToCart = async (req, res) => {
@@ -10,6 +12,13 @@ const addToCart = async (req, res) => {
     if (!productId) {
       return res.status(400).json({
         message: "Product ID is required",
+      });
+    }
+
+    // Validate Product ID format
+    if (!mongoose.Types.ObjectId.isValid(productId)) {
+      return res.status(400).json({
+        message: "Invalid Product ID",
       });
     }
 
@@ -35,6 +44,7 @@ const addToCart = async (req, res) => {
     });
 
     res.status(201).json(cartItem);
+
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -45,6 +55,14 @@ const addToCart = async (req, res) => {
 // PUT /cart/:id
 const updateCart = async (req, res) => {
   try {
+
+    // Validate Cart ID
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({
+        message: "Invalid Cart ID",
+      });
+    }
+
     const { quantity } = req.body;
 
     // Validate Quantity
@@ -67,6 +85,7 @@ const updateCart = async (req, res) => {
     }
 
     res.status(200).json(item);
+
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -77,6 +96,14 @@ const updateCart = async (req, res) => {
 // DELETE /cart/:id
 const deleteCart = async (req, res) => {
   try {
+
+    // Validate Cart ID
+    if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
+      return res.status(400).json({
+        message: "Invalid Cart ID",
+      });
+    }
+
     const item = await Cart.findByIdAndDelete(req.params.id);
 
     if (!item) {
@@ -88,6 +115,7 @@ const deleteCart = async (req, res) => {
     res.status(200).json({
       message: "Item removed from cart",
     });
+
   } catch (error) {
     res.status(500).json({
       message: error.message,
@@ -95,8 +123,4 @@ const deleteCart = async (req, res) => {
   }
 };
 
-module.exports = {
-  addToCart,
-  updateCart,
-  deleteCart,
-};
+module.exports = { addToCart, updateCart, deleteCart};
